fix: Refactor code

+++ src/controllers/Products/products.ts
@@ -1,19 +1,20 @@
-import {Request, Response} from 'express';
+import { Request, Response } from 'express';
 import prisma from '../../database/prismaClient';
-import {productsDto} from '../Products/products.dto';
+import { productsDto } from '../Products/products.dto';
 
 const createProduct = async (req: Request, res: Response) => {
- try {
-  const {nome, descricao, preco, estoque} : productsDto = req.body;
-  const newProduct = await prisma.produtos.create(
-    {
-      data: {
-        nome, 
-        descricao, 
-        preco, 
-        estoque, 
-        created_at: new Date()
-    }})
+  try {
+    const { nome, descricao, preco, estoque }: productsDto = req.body;
+    const newProduct = await prisma.produtos.create(
+      {
+        data: {
+          nome,
+          descricao,
+          preco,
+          estoque,
+          created_at: new Date()
+        }
+      })
 
     const resultProduct = {
       id: newProduct.id,
@@ -24,62 +25,83 @@ const createProduct = async (req: Request, res: Response) => {
       created_at: newProduct.created_at,
     };
     return res.status(200).json(resultProduct);
- } catch (error) {
-  console.log(error);
-  
-  return res.status(500).json({message: "Error interno do servidor"});
- }
+  } catch (error) {
+    return res.status(500).json({ message: "Error interno do servidor" });
+  }
 }
 
-const updateProduct = async (req: Request, res: Response) =>{
+const updateProduct = async (req: Request, res: Response) => {
   try {
-    const {id} = req.params;
-    const {nome, descricao, preco, estoque} : productsDto = req.body;
-    const existProduct = await prisma.produtos.findFirst({where: {id: parseInt(id)}})
+    const { id } = req.params;
+    const { nome, descricao, preco, estoque }: productsDto = req.body;
+    const existProduct = await prisma.produtos.findFirst({ where: { id: parseInt(id) } })
 
-    if(!existProduct){
-      return res.status(404).json({message: "Produto não encontrado"})
+    if (!existProduct) {
+      return res.status(404).json({ message: "Produto não encontrado" })
     }
 
     const updateProduct = await prisma.produtos.update({
-      where: {id: parseInt(id)},
-      data:{
+      where: { id: parseInt(id) },
+      data: {
         nome,
         descricao,
         preco,
         estoque: existProduct.estoque + estoque
       }
-  });
+    });
 
 
-   return res.status(200).json(updateProduct)
+    return res.status(200).json(updateProduct)
 
 
   } catch (error) {
-    return res.status(500).json({message: "Error interno do servidor"});
+    return res.status(500).json({ message: "Error interno do servidor" });
   }
 }
 
-const deleteProduct = async (req: Request, res: Response) =>{
- try {
-  const {id} = req.params;
-  const existProduct = await prisma.produtos.findFirst({where: {id: parseInt(id)}})
-  if(!existProduct){
-    return res.status(404).json({message: "Produto não encontrado"})
+const deleteProduct = async (req: Request, res: Response) => {
+  try {
+    const { id } = req.params;
+    const existProduct = await prisma.produtos.findFirst({ where: { id: parseInt(id) } })
+    if (!existProduct) {
+      return res.status(404).json({ message: "Produto não encontrado" })
+    }
+
+    const existOrder = await prisma.pedidos.findFirst({ where: { produto_id: parseInt(id) } })
+    if (existOrder) {
+      return res.status(404).json({ message: "Produto não pode ser deletado, pois existe pedido vinculado." })
+    }
+    await prisma.produtos.delete({ where: { id: parseInt(id) } })
+
+    return res.status(200).send();
+
+  } catch (error) {
+    return res.status(500).json({ message: "Error interno do servidor" });
   }
 
-  const deleteProduct  = await prisma.produtos.delete({where: {id:parseInt(id)}})
+};
 
-   return res.status(200).send();
+const showStock = async (req: Request, res: Response) => {
+  try {
+    const productStock = await prisma.produtos.findMany({
+      where: {
+        estoque: {
+          gt: 0
+        }
+      }
+    });
 
- } catch (error) {
-  
- }
-  
+    return res.status(200).json(productStock);
+  } catch (error) {
+    return res.status(500).json({ message: "Error interno do servidor" });
+
+  }
+
 }
 
 export default {
   createProduct,
   updateProduct,
-  deleteProduct
+  deleteProduct,
+  showStock
 }
